+++ satella/cassandra/future.py
@@ -6,14 +6,15 @@ from .common import ResponseFuture
 def wrap_future(future: ResponseFuture) -> Future:
     """
     Convert a Cassandra's future to a normal Python future.
-    The future is returned already
+    The returned future will be marked as running.
+
+    future is returned when it's already a Python future.
 
     :param future: cassandra future to wrap
     :return: a standard Python future
-    :raises TypeError: when a normal Python future is passed as future
     """
     if isinstance(future, Future):
-        raise TypeError('Tried to wrap an existing Future!')
+        return future
 
     fut = Future()
     fut.set_running_or_notify_cancel()