+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.15.7
 
 * deprecated `IteratorListAdapter`
+* fixed a bug that disabled json if ujson was not installed