+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-# v2.15.8
+# v2.16
 
+* changed semantics of `wrap_future`